api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "terceirizados-api"
version = "0.1.0"
description = "API de consulta de dados de terceirizados usando DuckDB"
authors = [
    { name = "Talissa Moura" }
]
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "flask>=3.0",
    "duckdb>=0.10",
    "google-cloud-storage>=2.16",
    "flasgger>=0.9"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
